filter preset lines by job
